+++ R/bed_window.r
@@ -1,7 +1,10 @@
 #' Identify intervals within a specified distance.
 #'
 #' @param x [ivl_df]
-#' @param y  [ivl_df]
+#' @param y [ivl_df], or a path or URL to a bigWig (`.bw`) or bigBed (`.bb`)
+#'   file. When a file is supplied, only the windowed regions around `x` are
+#'   read from it (local files and `http(s)://` URLs are both supported),
+#'   avoiding the cost of loading the entire file.
 #' @param ... params for bed_slop and bed_intersect
 #' @inheritParams bed_slop
 #' @inheritParams bed_intersect
@@ -71,7 +74,10 @@ bed_window <- function(x, y, genome, ...) {
   check_required(genome)
 
   x <- check_interval(x)
-  y <- check_interval(y)
+  # a bigWig/bigBed `y` is resolved below, once the window has been applied
+  if (!is_bigwig_path(y)) {
+    y <- check_interval(y)
+  }
   genome <- check_genome(genome)
 
   x <- mutate(x, .start = .data[["start"]], .end = .data[["end"]])
@@ -92,6 +98,11 @@ bed_window <- function(x, y, genome, ...) {
     bed_slop,
     c(list("x" = x, "genome" = genome), slop_args)
   )
+
+  # for a bigWig/bigBed `y`, query only the windowed (slopped) x regions
+  if (is_bigwig_path(y)) {
+    y <- read_bigwig_regions(slop_x, y)
+  }
 
   # pass new list of args to bed_intersect
   # TODO: revisit when min_overlap default changes to 1L